feat: Comprehensive project refactor and enhancement
This commit introduces a wide range of improvements across the entire AI Sensei project, enhancing maintainability, performance, and robustness.

Key changes include:

1.  **Project Structure Cleanup:**
    - Removed the redundant nested `functions/functions` directory.
    - Regenerated a clean `package-lock.json` for the functions directory.

2.  **Backend Code Enhancements:**
    - Aligned module imports in `functions/src/index.ts` with TypeScript best practices by removing the `.js` extension.
    - Unified the CORS policy across all Cloud Functions for consistency and security.
    - Improved error logging in `functions/src/gemini-api.ts` to include the original error object for more effective debugging.

3.  **Frontend Refactoring and Robustness:**
    - Refactored the monolithic `public/js/main.js` into a modular architecture with `app.js`, `auth.js`, `professor.js`, and `student.js`.
    - Corrected the application's entry point to `firebase-init.js` in `index.html` to ensure proper initialization.
    - Implemented robust YouTube URL validation in the professor's video submission form with clear user feedback.

4.  **Dependency Management:**
    - Moved `firebase-admin` and `firebase-tools` from `dependencies` to `devDependencies` in the root `package.json` as they are not client-side dependencies.

5.  **CI/CD Optimization:**
    - Implemented caching for `node_modules` in the GitHub Actions workflow for both root and `functions` directories to accelerate deployment times.
