+++ .github/workflows/deploy.yml
@@ -17,18 +17,33 @@ jobs:
         with:
           node-version: '22'
 
+      - name: Cache root node_modules
+        id: cache-root-modules
+        uses: actions/cache@v3
+        with:
+          path: '**/node_modules'
+          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
+
       - name: Install root dependencies
+        if: steps.cache-root-modules.outputs.cache-hit != 'true'
         run: npm install
 
       - name: Clean previous build artifacts
         run: rm -rf functions/lib
 
+      - name: Cache functions node_modules
+        id: cache-functions-modules
+        uses: actions/cache@v3
+        with:
+          path: functions/node_modules
+          key: ${{ runner.os }}-functions-node-${{ hashFiles('functions/package-lock.json') }}
+
       - name: Install Functions dependencies and build
+        if: steps.cache-functions-modules.outputs.cache-hit != 'true'
         run: |
-          cd functions
-          npm ci
-          npm run lint
-          npm run build
+          npm install --prefix functions
+          npm run lint --prefix functions
+          npm run build --prefix functions
 
       - name: Authenticate to Google Cloud
         id: 'auth'
