+++ package.json
@@ -12,11 +12,11 @@
   "type": "commonjs",
   "dependencies": {
     "firebase": "^12.3.0",
-    "firebase-admin": "^13.5.0",
-    "firebase-tools": "^14.17.0",
     "uuid": "^13.0.0"
   },
   "devDependencies": {
-    "@types/uuid": "^10.0.0"
+    "@types/uuid": "^10.0.0",
+    "firebase-admin": "^13.5.0",
+    "firebase-tools": "^14.17.0"
   }
 }
