+++ public/index.html
@@ -199,7 +199,7 @@
         </div>
     </div>
 
-    <!-- The main application script will now handle all logic -->
-    <script type="module" src="/js/main.js"></script>
+    <!-- Firebase Init is the entry point that bootstraps the entire application -->
+    <script type="module" src="/js/firebase-init.js"></script>
 </body>
 </html>
